+++ tests/server/blueprints/api_v1/test_views_add.py
@@ -1,47 +1,83 @@
 from cgbeacon2.resources import test_snv_vcf_path
 import json
 
-HEADERS = {"Content-type": "application/json", "Accept": "application/json"}
+
+def test_add_variants_no_auth_token(mock_app, api_req_headers):
+    """Test receiving an add request with no X-Auth-Token header key"""
+    api_req_headers.pop("X-Auth-Token")
+    data = dict(vcf_path="path/to/vcf", assemblyId="GRCh37")
+    # When a POST add request is missing the token in the header
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
+    # then it should return not authorized
+    assert response.status_code == 403
+    resp_data = json.loads(response.data)
+    assert resp_data["message"] == "Invalid auth token error"
 
 
-def test_variants_add_no_dataset(mock_app):
+def test_add_variants_wrong_auth_token(mock_app, database, api_user, api_req_headers):
+    """Test receiving an add request with an X-Auth-Token header key not registered in database"""
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
+    # WHEN an add request with the wrong token is sent
+    api_req_headers["X-Auth-Token"] = "FOO"
+    data = dict(vcf_path="path/to/vcf", assemblyId="GRCh37")
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
+    # then it should return not authorized
+    assert response.status_code == 403
+    resp_data = json.loads(response.data)
+    assert resp_data["message"] == "Invalid auth token error"
+
+
+def test_variants_add_no_dataset(mock_app, api_user, database, api_req_headers):
     """Test receiving a variant add request missing one of the required params"""
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
+
     data = dict(vcf_path="path/to/vcf", assemblyId="GRCh37")
     # When a POST add request is missing dataset id param:
-    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
     # Then it should return error 422 (Unprocessable Entity)
     assert response.status_code == 422
     resp_data = json.loads(response.data)
     assert resp_data["message"] == "'dataset_id' is a required property"
 
 
-def test_variants_add_no_vcf_path(mock_app):
+def test_variants_add_no_vcf_path(mock_app, api_user, database, api_req_headers):
     """Test receiving a variant add request missing one of the required params"""
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
+
     data = dict(dataset_id="test_id", assemblyId="GRCh37")
     # When a POST add request is missing dataset path to vcf file
-    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
     # Then it should return error 422 (Unprocessable Entity)
     assert response.status_code == 422
     resp_data = json.loads(response.data)
     assert resp_data["message"] == "'vcf_path' is a required property"
 
 
-def test_variants_add_wrong_assembly(mock_app):
+def test_variants_add_wrong_assembly(mock_app, api_user, database, api_req_headers):
     """Test receiving a variant add request with non-valid genome assembly"""
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
+
     data = dict(dataset_id="test_id", vcf_path="path/to/vcf", assemblyId="FOO")
     # When a POST add request is sent with a non valid assembly id
-    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
     # Then it should return error 422 (Unprocessable Entity)
     assert response.status_code == 422
     resp_data = json.loads(response.data)
     assert resp_data["message"] == "'FOO' is not one of ['GRCh37', 'GRCh38']"
 
 
-def test_variants_add_wrong_dataset(mock_app):
+def test_variants_add_wrong_dataset(mock_app, api_user, database, api_req_headers):
     """Test receiving a variant add request with non-valid dataset id"""
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
+
     data = dict(dataset_id="FOO", vcf_path="path/to/vcf", assemblyId="GRCh37")
     # When a POST add request is sent with a non valid assembly id
-    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
     # Then it should return error
     assert response.status_code == 422
     # With message that dataset could not be found
@@ -49,15 +85,18 @@ def test_variants_add_wrong_dataset(mock_app):
     assert data["message"] == "Invalid request. Please specify a valid dataset ID"
 
 
-def test_variants_add_invalid_vcf_path(mock_app, public_dataset, database):
+def test_variants_add_invalid_vcf_path(
+    mock_app, public_dataset, database, api_user, api_req_headers
+):
     """Test receiving a variant add request with non-valid vcf path"""
-
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
     # GIVEN a database containing a public dataset
     database["dataset"].insert_one(public_dataset)
 
     data = dict(dataset_id=public_dataset["_id"], vcf_path="path/to/vcf", assemblyId="GRCh37")
     # When a POST add request is sent with a non valid assembly id
-    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
     # Then it should return error
     assert response.status_code == 422
     # With message that VCF path is not valid
@@ -65,9 +104,12 @@ def test_variants_add_invalid_vcf_path(mock_app, public_dataset, database):
     assert data["message"] == "Error extracting info from VCF file, please check path to VCF"
 
 
-def test_variants_add_invalid_samples(mock_app, public_dataset, database):
+def test_variants_add_invalid_samples(
+    mock_app, public_dataset, database, api_user, api_req_headers
+):
     """Test receiving a variant add request with non-valid samples (samples not in provided VCF file)"""
-
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
     # GIVEN a database containing a public dataset
     database["dataset"].insert_one(public_dataset)
 
@@ -78,7 +120,7 @@ def test_variants_add_invalid_samples(mock_app, public_dataset, database):
         samples=["FOO", "BAR"],
     )
     # When a POST add request is sent with non-valid samples
-    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
     # Then it should return error
     assert response.status_code == 422
     # With message that VCF files doesn't contain those samples
@@ -86,9 +128,12 @@ def test_variants_add_invalid_samples(mock_app, public_dataset, database):
     assert "One or more provided samples were not found in VCF" in data["message"]
 
 
-def test_variants_add_invalid_gene_list(mock_app, public_dataset, database):
+def test_variants_add_invalid_gene_list(
+    mock_app, public_dataset, database, api_user, api_req_headers
+):
     """Test receiving a variant add request with non-valid genes object"""
-
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
     # GIVEN a database containing a public dataset
     database["dataset"].insert_one(public_dataset)
 
@@ -100,7 +145,7 @@ def test_variants_add_invalid_gene_list(mock_app, public_dataset, database):
         genes={"ids": [17284]},
     )
     # When a POST add request is sent with non-valid genes object (missing id_type for instance)
-    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
     # Then it should return error
     assert response.status_code == 422
     # With message that missing info should be provided
@@ -108,9 +153,12 @@ def test_variants_add_invalid_gene_list(mock_app, public_dataset, database):
     assert "Please provide id_type (HGNC or Ensembl) for the given list of genes" in data["message"]
 
 
-def test_variants_add_hgnc_genes(mock_app, public_dataset, database, test_gene):
+def test_variants_add_hgnc_genes(
+    mock_app, public_dataset, database, test_gene, api_user, api_req_headers
+):
     """Test receiving a request to add variants with valid parameters, hgnc genes"""
-
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
     # GIVEN a database containing a public dataset
     database["dataset"].insert_one(public_dataset)
 
@@ -129,7 +177,7 @@ def test_variants_add_hgnc_genes(mock_app, public_dataset, database, test_gene):
         "samples": samples,
         "genes": {"ids": [17284], "id_type": "HGNC"},
     }
-    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add", json=data, headers=api_req_headers)
     # Then it should return a success response
     assert response.status_code == 200
     resp_data = json.loads(response.data)
@@ -142,9 +190,12 @@ def test_variants_add_hgnc_genes(mock_app, public_dataset, database, test_gene):
     assert updated_dataset["samples"] == samples
 
 
-def test_variants_add_ensembl_genes(mock_app, public_dataset, database, test_gene):
+def test_variants_add_ensembl_genes(
+    mock_app, public_dataset, database, test_gene, api_user, api_req_headers
+):
     """Test receiving a request to add variants with valid parameters, hgnc genes"""
-
+    # GIVEN an authorized API user
+    database["user"].insert_one(api_user)
     # GIVEN a database containing a public dataset
     database["dataset"].insert_one(public_dataset)
 
@@ -160,7 +211,7 @@ def test_variants_add_ensembl_genes(mock_app, public_dataset, database, test_gen
         "samples": samples,
         "genes": {"ids": ["ENSG00000128513"], "id_type": "Ensembl"},
     }
-    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=HEADERS)
+    response = mock_app.test_client().post("/apiv1.0/add?", json=data, headers=api_req_headers)
     # Then it should return a success response
     assert response.status_code == 200
     resp_data = json.loads(response.data)